[12.0][MIG] mrp_multi_level: * You can know integrate with different forecasting mechanisms.   As a consequence integration with stock_demand_estimate is moved   to a new module (mrp_multi_level_estimate). * As agreed by the authors, the module is re-licensed to LGPL-3.
