+++ mrp_multi_level/demo/initial_on_hand_demo.xml
@@ -6,26 +6,26 @@
     </record>
     <record id="stock_inventory_line_1" model="stock.inventory.line">
         <field name="product_id" ref="product_product_pp_1"/>
-        <field name="product_uom_id" ref="product.product_uom_unit"/>
+        <field name="product_uom_id" ref="uom.product_uom_unit"/>
         <field name="inventory_id" ref="stock_inventory_mrp_example"/>
         <field name="product_qty">10</field>
         <field name="location_id" ref="stock.stock_location_stock"/>
     </record>
     <record id="stock_inventory_line_2" model="stock.inventory.line">
         <field name="product_id" ref="product_product_pp_2"/>
-        <field name="product_uom_id" ref="product.product_uom_unit"/>
+        <field name="product_uom_id" ref="uom.product_uom_unit"/>
         <field name="inventory_id" ref="stock_inventory_mrp_example"/>
         <field name="product_qty">20</field>
         <field name="location_id" ref="stock.stock_location_stock"/>
     </record>
     <record id="stock_inventory_line_3" model="stock.inventory.line">
         <field name="product_id" ref="product_product_sf_2"/>
-        <field name="product_uom_id" ref="product.product_uom_unit"/>
+        <field name="product_uom_id" ref="uom.product_uom_unit"/>
         <field name="inventory_id" ref="stock_inventory_mrp_example"/>
         <field name="product_qty">15</field>
         <field name="location_id" ref="stock.stock_location_stock"/>
     </record>
-    <function model="stock.inventory" name="action_done">
+    <function model="stock.inventory" name="action_validate">
         <function eval="[[('id', '=', ref('stock_inventory_mrp_example'))]]" model="stock.inventory" name="search"/>
     </function>
