+++ mrp_multi_level/demo/product_product_demo.xml
@@ -4,8 +4,8 @@
         <field name="name">FP-1</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
         <field name="produce_delay">2</field>
         <field name="route_ids" eval="[(6, 0, [ref('mrp.route_warehouse0_manufacture')])]"/>
     </record>
@@ -14,8 +14,8 @@
         <field name="name">FP-2</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
         <field name="produce_delay">1</field>
         <field name="route_ids" eval="[(6, 0, [ref('mrp.route_warehouse0_manufacture')])]"/>
     </record>
@@ -24,8 +24,8 @@
         <field name="name">SF-1</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
         <field name="produce_delay">1</field>
         <field name="route_ids" eval="[(6, 0, [ref('mrp.route_warehouse0_manufacture')])]"/>
     </record>
@@ -33,8 +33,8 @@
         <field name="name">SF-2</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
         <field name="produce_delay">3</field>
         <field name="route_ids" eval="[(6, 0, [ref('mrp.route_warehouse0_manufacture')])]"/>
     </record>
@@ -43,17 +43,17 @@
         <field name="name">PP-1</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
-        <field name="route_ids" eval="[(6, 0, [ref('purchase.route_warehouse0_buy')])]"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
+        <field name="route_ids" eval="[(6, 0, [ref('purchase_stock.route_warehouse0_buy')])]"/>
     </record>
 
     <record id="product_product_pp_2" model="product.product">
         <field name="name">PP-2</field>
         <field name="categ_id" ref="product_category_mrp"/>
         <field name="type">product</field>
-        <field name="uom_id" ref="product.product_uom_unit"/>
-        <field name="uom_po_id" ref="product.product_uom_unit"/>
-        <field name="route_ids" eval="[(6, 0, [ref('purchase.route_warehouse0_buy')])]"/>
+        <field name="uom_id" ref="uom.product_uom_unit"/>
+        <field name="uom_po_id" ref="uom.product_uom_unit"/>
+        <field name="route_ids" eval="[(6, 0, [ref('purchase_stock.route_warehouse0_buy')])]"/>
     </record>
 </odoo>
