+++ mrp_multi_level/models/product_product.py
@@ -1,6 +1,6 @@
 # Copyright 2016 Ucamco - Wim Audenaert <wim.audenaert@ucamco.com>
 # Copyright 2016-18 Eficent Business and IT Consulting Services S.L.
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 import ast
 from odoo import api, fields, models
