+++ mrp_multi_level/models/product_template.py
@@ -1,5 +1,5 @@
 # Copyright 2018 Eficent Business and IT Consulting Services S.L.
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 import ast
 from odoo import api, fields, models
