+++ mrp_multi_level/models/stock_location.py
@@ -1,7 +1,7 @@
 # © 2016 Ucamco - Wim Audenaert <wim.audenaert@ucamco.com>
 # © 2016 Eficent Business and IT Consulting Services S.L.
 # - Jordi Ballester Alomar <jordi.ballester@eficent.com>
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 from odoo import fields, models
