+++ mrp_multi_level/tests/__init__.py
@@ -1,2 +1,1 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import test_mrp_multi_level
