+++ mrp_multi_level/views/product_mrp_area_views.xml
@@ -51,7 +51,6 @@
                             <field name="mrp_exclude"/>
                             <field name="mrp_verified"/>
                             <field name="mrp_nbr_days"/>
-                            <field name="group_estimate_days"/>
                             <!--hide delays for now-->
                             <field name="mrp_transit_delay" invisible="1"/>
                             <field name="mrp_inspection_delay" invisible="1"/>
