+++ mrp_multi_level/wizards/mrp_inventory_procure_views.xml
@@ -21,7 +21,7 @@
                             <field name="location_id" groups="stock.group_stock_multi_locations" readonly="1"/>
                             <field name="product_id" readonly="1"/>
                             <field name="qty"/>
-                            <field name="uom_id" groups="product.group_uom"/>
+                            <field name="uom_id" groups="uom.group_uom"/>
                             <field name="date_planned"/>
                         </tree>
                     </field>
